list plat et afficher description plat

+++ src/main/java/epsi/model/Menu.java
@@ -20,7 +20,7 @@ import org.hibernate.annotations.Parameter;
 @Table(name = "menu", catalog = "bdd_lpdc")
 public class Menu implements java.io.Serializable {
 
-	private int idProduit;
+	private long idProduit;
 	private Produit produit;
 	private int nbPersonne;
 	private Float prix;
@@ -48,11 +48,11 @@ public class Menu implements java.io.Serializable {
 	@Id
 	@GeneratedValue(generator = "generator")
 	@Column(name = "id_Produit", unique = true, nullable = false)
-	public int getIdProduit() {
+	public long getIdProduit() {
 		return this.idProduit;
 	}
 
-	public void setIdProduit(int idProduit) {
+	public void setIdProduit(long idProduit) {
 		this.idProduit = idProduit;
 	}
